requirements.txt
Pillow
matplotlib
pandas
exif